+++ Cargo.toml
@@ -34,7 +34,7 @@ members = [
   "primitives/bitfield/rle",
   "primitives/block",
   "primitives/bytes",
-  "primitives/cid-ext",
+  "primitives/fc",
   "primitives/hash",
   "primitives/message",
   "primitives/piece",
