+++ primitives/piece/src/piece.rs
@@ -1,12 +1,8 @@
 // Copyright 2019-2020 PolkaX Authors. Licensed under GPL-3.0.
-use std::convert::TryInto;
 
 use cid::Cid;
-use filecoin_proofs::types;
 use minicbor::{decode, encode, Decoder, Encoder};
 use serde::{Deserialize, Serialize};
-
-use cid_ext::comm;
 
 use crate::size::PaddedPieceSize;
 
@@ -39,29 +35,4 @@ impl<'b> decode::Decode<'b> for PieceInfo {
             piece_cid: d.decode::<Cid>()?,
         })
     }
-}
-
-impl TryInto<types::PieceInfo> for PieceInfo {
-    type Error = comm::CommCidErr;
-
-    fn try_into(self) -> Result<types::PieceInfo, Self::Error> {
-        let unpadded = self.size.unpadded();
-        let commitment = comm::cid_to_piece_commitment_v1(&self.piece_cid)?;
-        Ok(types::PieceInfo {
-            commitment,
-            size: unpadded.into(),
-        })
-    }
-}
-
-/// convert piece info list to filecoin proof pieceinfo list
-pub fn convert_pieceinfos(
-    pieceinfos: Vec<PieceInfo>,
-) -> Result<Vec<types::PieceInfo>, comm::CommCidErr> {
-    let mut v = Vec::with_capacity(pieceinfos.len());
-    for info in pieceinfos {
-        let p = info.try_into()?;
-        v.push(p);
-    }
-    Ok(v)
 }
