+++ primitives/sector/Cargo.toml
@@ -15,7 +15,6 @@ serde_repr = "0.1"
 plum_bigint = { path = "../bigint" }
 plum_bytes = { path = "../bytes" }
 plum_types = { path = "../types" }
-plum_address = { path = "../address" }
 
 [dev-dependencies]
 serde_json = "1.0"
