+++ primitives/sector/src/lib.rs
@@ -11,8 +11,8 @@ mod sector;
 pub use self::posting::{PoStProof, WindowPoStVerifyInfo, WinningPoStVerifyInfo};
 pub use self::sealing::{OnChainSealVerifyInfo, SealVerifyInfo};
 pub use self::sector::{
-    readable_sector_size, to_prove_id, RegisteredProof, SectorId, SectorInfo, SectorNumber,
-    SectorQuality, SectorSize, SpaceTime, StoragePower,
+    readable_sector_size, RegisteredProof, SectorId, SectorInfo, SectorNumber, SectorQuality,
+    SectorSize, SpaceTime, StoragePower,
 };
 
 #[cfg(test)]
