+++ primitives/types/Cargo.toml
@@ -12,5 +12,6 @@ minicbor = { version = "0.4", features = ["std"] }
 serde = { version = "1.0", features = ["derive"] }
 
 # plum
+plum_address = { path = "../address" }
 plum_bigint = { path = "../bigint" }
 plum_hash = { path = "../hash" }
